restructure caused path issues. fixed
