+++ client/src/controllers/Profile/Profile.js
@@ -4,7 +4,7 @@ import './Profile.scss'
 
 import YourStuff from './YourStuff/YourStuff'
 // import OthersStuff from './OthersStuff/OthersStuff'
-import HomeButton from '../../components/HomeButton/HomeButton';
+import GoHomeButton from '../../components/ButtonTypes/GoHomeButton/GoHomeButton'
 
 const Profile = props => {
     const [yourStuff, isYourStuff] = useState(true)
@@ -12,7 +12,7 @@ const Profile = props => {
 
     return (
         <div className='profile'>
-            <HomeButton />
+            <GoHomeButton />
 
             <h1 className='pageTitle'>{user?user.username:'Not Logged In'}</h1>
